src/test/java/com/devoxx/genie/service/DocumentServiceTest.java
package com.devoxx.genie.service;

import com.devoxx.genie.domain.Content;
import com.devoxx.genie.domain.EmbeddingModelReference;
import com.devoxx.genie.domain.User;
import com.devoxx.genie.repository.ContentRepository;
import com.devoxx.genie.repository.EmbeddingModelReferenceRepository;
import com.devoxx.genie.repository.UserRepository;
import com.devoxx.genie.service.dto.DocumentDTO;
import com.devoxx.genie.service.dto.VectorDocumentDTO;
import dev.langchain4j.data.segment.TextSegment;
import dev.langchain4j.store.embedding.EmbeddingMatch;
import org.junit.Ignore;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.jdbc.core.JdbcTemplate;

import java.util.ArrayList;
import java.util.List;
import java.util.UUID;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;

@GenieServiceTest
class DocumentServiceTest {

    @Autowired
    DocumentService documentService;

    @Autowired
    EmbeddingModelReferenceRepository embeddingModelReferenceRepository;

    @Autowired
    ContentRepository contentRepository;

    @Autowired
    UserRepository userRepository;

    @Autowired
    JdbcTemplate jdbcTemplate;

    @Test
    void saveDocument() {
        List<String> chunks = new ArrayList<>();
        chunks.add("Hello world");

        Content content = new Content();
        content.setSource("Hello world");

        Content savedContent = contentRepository.save(content);
        assertThat(savedContent).isNotNull();

        User user = userRepository.findAll().getFirst();

        EmbeddingModelReference embedModel = embeddingModelReferenceRepository.findAll().getFirst();

        List<String> ids = documentService.save(user.getId(), savedContent.getId(), chunks, embedModel.getId());

        assertThat(ids).isNotNull();
        assertThat(ids.size()).isEqualTo(1);
    }

    private int countRecords(int dimSize) {
        String sql = "SELECT COUNT(*) FROM genie_document_%d".formatted(dimSize);
        return jdbcTemplate.query(sql, (rs, rowNum) -> rs.getInt(1)).getFirst();
    }

    @Test
    void updateDocument() {
        User firstUser = userRepository.findAll().getFirst();
        VectorDocumentDTO vectorDocument = new VectorDocumentDTO(UUID.randomUUID(), "Hello world");

        documentService.update(firstUser.getId(), 384, vectorDocument);
    }

    @Test
    void getByUserId() {
        Page<DocumentDTO> byUserId = documentService.getByUserIdAndDimension(3L, 512, Pageable.unpaged());
        assertThat(byUserId).isNotNull();
    }

    @Test
    void findAllContentEmbeddingsForUserId() {
        float[][] embeddings =
            documentService.findAllContentEmbeddingsForUserId(384, 3L, 3L);
        assertThat(embeddings).isNotNull();
    }

    // Not really sure why this is failing...
    @Ignore
    void getByDocumentId() {
        List<String> chunks = new ArrayList<>();
        chunks.add("Hello world");

        Content content = new Content();
        content.setSource("Hello world");

        Content savedContent = contentRepository.save(content);
        assertThat(savedContent).isNotNull();

        User user = userRepository.findAll().getFirst();

        EmbeddingModelReference embedModel = embeddingModelReferenceRepository.findAll().getFirst();

        List<String> savedIds = documentService.save(user.getId(), savedContent.getId(), chunks, embedModel.getId());
        assertThat(savedIds).isNotNull();
        UUID uuid = UUID.fromString(savedIds.getFirst());
        documentService.getDocumentByUserIdDimensionAndDocumentUUID(uuid, embedModel.getDimSize())
            .ifPresentOrElse(document -> {
                assertThat(document).isNotNull();
                assertThat(document.getId()).isEqualTo(uuid);
            }, () -> {
                throw new RuntimeException("Document not found");
            });
    }

    @Test
    void findRelevantDocuments() {
        List<EmbeddingMatch<TextSegment>> relevantDocuments =
            documentService.findRelevantDocuments(3L, 1L,
            "Give me all docs related to Devoxx", 3, 0.6);
        assertThat(relevantDocuments).isNotNull();
    }

    @Test
    void delete() {
        documentService.delete(384, UUID.randomUUID());
    }

    @Test
    void deleteAllByUserId() {
        documentService.deleteAllByDimensionForUserId(384, 3L);
    }

    @Test
    void testFindAllByUserId() {

        documentService.getByUserIdAndDimension(3L, 512, Pageable.unpaged());
    }
}
